Rename TableView to ListView
