+++ packages/lib-utils/src/components/list-view/index.ts
@@ -1,0 +1,1 @@
+export { default as ListView, FilterItem, ListViewProps } from './ListView';
