--- packages/lib-utils/src/components/table-view/table-view.css
@@ -1,11 +1,0 @@
-.pf-c-search-input.dps-table-view__search {
-  margin-right: var(--pf-global--spacer--sm);
-}
-
-.pf-c-toolbar__item.dps-table-view__top-pagination {
-  margin-left: auto;
-}
-
-.pf-c-toolbar__content.dps-table-view__filters {
-  margin-top: calc(-1 * var(--pf-global--spacer--lg));
-}
